Work on TabBar
Currently. TabBar is now dynamically creating a new tab for every new
NuiTerm Terminal Window that is created.
Each Tab will update thei highlight depending on which tab is in focus.

I've also added a few more keymaps
  - NextTerm[ <leader>tl ]
  - PrevTerm[ <leader>th ]
  - NewTerm[ <leader>tn ]

Next, I will be implementing <RightClick> on Tab to change NuiTerm
Focus. And I'll be expanding on the TabDesign, Tab Naming.

I would like to look into Terminal State, in regards to Neovim.
Would it be possible to open up a terminal in a specfic state?