NuiTermDelete Now working correctly.
Before, MainWindow.DeleteTerm was failing to correctly a NuiTerm
instance correctly. Somehow, the buffer for term_id+2 was being
copied into term_id+1..

After rewriting the logic for this function. NuiTerminals are now
removed as expected. Though.. it may not be the most performant..
But, it's not like users will have 100+ terminal instances running at
any given moment.