Terminal Width Commands
Added support to win_config for width to be either a set width or a
percentage defined with a numerator and denominator
i.e.,
  - width = 80 (colunns)
  - width = 1/1 (full)
  - width = 1/3 (33% terminal width)
  - width = 5/6
  - etc..
When a non-full sized Width is picked. The Terminal window is perfectly
centered on screen

This currently only works on startup. I still need to implement
this change into MainWindow.OnResize
And I plan to add UserCommands for Expanding and Shrinking the width
of the terminal.