Made an important UI/UX Update -- Automatic placement of the Terminal Cursor.
After taking a break from development of NuiTerm for a couple days,
whilst I worked on other projects and using NuiTerm in a real world
setting. I came across a few nuiances that needed to worked on. One
of them being the cursor's placement when you reopen NuiTerm.

-- Issue:
 |  Before, if you had a long lived terminal window session, when reopenning
 |  NuiTerm, the cursor would be placed at line #0. So, for situations where
 |  I only wanted to read the most recent output. I had to put NuiTerm into
 |  TermMode to move the cursor to the bottom of the window..

-- Resolution:
 |  Created a new TermWindow function, TermWindow:MoveToLastLine: In this
 |  method, we call upon vim.api.nvim_buf_line_count to obtain the total
 |  buffer rows.
 |  Next, we then iterate in reverse starting at 'line_count'. On each
 |  iteration we extract the contents of each buffer line via
 |   - vim.nvim_buf_get_lines(self.bufnr, i-1, false)[1]
 |  then we test to see if both 'line' is not nil and if 'line' matchs
 |  "%s"(i.e., anything but empty space).
 |  If so, we then return the current iteration index.
 |  Using the index by passing the value into
 |  vim.api.nvim_win_set_cursor(self.winid, {line_count, 0})

Though, we could have simply passed the value returned by
nvim_buf_line_count. But, this caused a somewhat similar issue. When we
open a fresh terminal instance, the cursor is moved to the absolute
bottom of the NuiTerm window. Where it's more natural for the cursor to
be on the current terminal input line.
One positive surprise from this update. Before, I was unable to get the
cursor to move off of coloumn 0. Which did not lool natural at all.
I original tried to pass {4, 4} to nvim_win_set_cursor, but this only
moved the cursor to row 4 and NOT column 4.
Now, the cursor is being place exactly where it should be. Which is
honestly odd.